refactor: tsc build for everything
