+++ package.json
@@ -9,9 +9,11 @@
   "scripts": {
     "test": "pnpm run test",
     "dev": "pnpm run dev",
-    "build-renderer": "pnpm --filter blast-renderer run build ",
-    "start-renderer:dev": "pnpm --filter blast-renderer run start:dev ",
-    "start-renderer": "pnpm --filter blast-renderer run start ",
+    "build-renderer": "pnpm --filter blast-renderer run build",
+    "start-renderer:dev": "pnpm --filter blast-renderer run start:dev",
+    "start-renderer": "pnpm --filter blast-renderer run start",
+    "build-api": "pnpm --filter blast-api run build",
+    "build-utils": "pnpm --filter blast-utils run build",
     "fmt": "prettier --write .",
     "lint": "eslint ."
   },
@@ -25,5 +27,33 @@
   "engines": {
     "node": ">=18.0.0",
     "pnpm": ">=6"
+  },
+  "devDependencies": {
+    "@types/debug": "^4.1.7",
+    "@types/jest": "^29.2.4",
+    "@types/node": "^18.8.3",
+    "@types/webpack": "^5.28.0",
+    "@typescript-eslint/eslint-plugin": "^5.46.1",
+    "@typescript-eslint/parser": "^5.47.0",
+    "cross-env": "^7.0.3",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-import-resolver-alias": "^1.1.2",
+    "eslint-import-resolver-typescript": "^3.5.2",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react": "^7.31.11",
+    "eslint": "^8.30.0",
+    "jest": "^29.3.1",
+    "nodemon": "^2.0.20",
+    "prettier": "2.8.1",
+    "ts-loader": "^9.4.2",
+    "ts-node": "^10.9.1",
+    "tslib": "^2.4.1",
+    "typescript": "^4.9.4",
+    "webpack-cli": "^5.0.1",
+    "webpack": "5"
+  },
+  "dependencies": {
+    "debug": "^4.3.4"
   }
 }
