+++ packages/blast-api/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "blast-api",
+  "version": "1.0.0",
+  "description": "",
+  "scripts": {
+    "test": "jest",
+    "build": "tsc"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@raycast/api": "^1.45.0",
+    "blast-renderer": "workspace:*",
+    "blast-utils": "workspace:*",
+    "fs-extra": "^11.1.0",
+    "react": "^18.2.0"
+  },
+  "devDependencies": {
+    "@types/fs-extra": "^9.0.13",
+    "@types/node": "^18.8.3",
+    "rpc-websockets": "^7.5.0"
+  },
+  "jest": {
+    "testEnvironment": "node",
+    "verbose": true
+  }
+}
