+++ packages/blast-api/src/Action/index.tsx
@@ -1,9 +1,8 @@
-import { Action as RaycastAction } from "raycast-original";
-
+import type { Action as RaycastAction } from "@raycast/api";
+import { ElementTypes } from "blast-renderer";
+import { createDebug } from "blast-utils";
 import { useCallback, useEffect, useId, useMemo } from "react";
 
-import * as ElementTypes from "../../renderer/elements/types";
-import { createDebug } from "../../utils/debug";
 import { useFormContext } from "../Form";
 import { useWsServer } from "../internal/WsServerProvider";
 import { useNavigation } from "../Navigation";
