+++ packages/blast-api/src/ActionPanel/index.tsx
@@ -1,6 +1,5 @@
-import { ActionPanel as RActionPanel } from "raycast-original";
-
-import * as ElementTypes from "../../renderer/elements/types";
+import type { ActionPanel as RActionPanel } from "@raycast/api";
+import { ElementTypes } from "blast-renderer";
 
 import { Section } from "./Section";
