+++ packages/blast-api/src/Detail/index.tsx
@@ -1,6 +1,5 @@
-import { Detail as RDetail } from "raycast-original";
-
-import * as ElementTypes from "../../renderer/elements/types";
+import { Detail as RDetail } from "@raycast/api";
+import { ElementTypes } from "blast-renderer";
 
 type DetailPropKeys = (keyof RDetail.Props)[];
 const serializedKeys: DetailPropKeys = ["isLoading", "markdown", "navigationTitle"];
