+++ packages/blast-api/src/Form/TextField.tsx
@@ -1,9 +1,8 @@
-import { Form } from "raycast-original";
-
+import { Form } from "@raycast/api";
+import { ElementTypes } from "blast-renderer";
+import { createDebug } from "blast-utils";
 import { forwardRef, useEffect, useId, useImperativeHandle, useMemo, useState } from "react";
 
-import * as ElementTypes from "../../renderer/elements/types";
-import { createDebug } from "../../utils/debug";
 import { useWsServer } from "../internal/WsServerProvider";
 
 import { useFormContext } from ".";
