+++ packages/blast-api/src/Form/index.tsx
@@ -1,9 +1,7 @@
-import { Form as RForm } from "raycast-original";
-
+import { Form as RForm } from "@raycast/api";
+import { ElementTypes } from "blast-renderer";
+import { createDebug } from "blast-utils";
 import { createContext, useCallback, useContext, useState } from "react";
-
-import * as ElementTypes from "../../renderer/elements/types";
-import { createDebug } from "../../utils/debug";
 
 import { TextField } from "./TextField";
