+++ packages/blast-api/src/List/index.tsx
@@ -1,6 +1,5 @@
-import { List as RList } from "raycast-original";
-
-import * as ElementTypes from "../../renderer/elements/types";
+import { List as RList } from "@raycast/api";
+import { ElementTypes } from "blast-renderer";
 
 import { Dropdown } from "./Dropdown";
 import { EmptyView } from "./EmptyView";
