+++ packages/blast-renderer/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "jest",
     "dev": "webpack --config webpack.config.ts --watch --progress",
-    "build": "webpack --config webpack.config.ts --progress",
+    "build": "tsc",
     "start:dev": "cross-env DEBUG=blast* nodemon dist/bundle.js",
     "start": "node dist/bundle.js",
     "fmt": "prettier --write .",
@@ -17,7 +17,6 @@
   "license": "ISC",
   "dependencies": {
     "@raycast/api": "^1.45.0",
-    "debug": "^4.3.4",
     "fs-extra": "^11.1.0",
     "global": "^4.4.0",
     "lodash": "^4.17.21",
@@ -25,38 +24,14 @@
     "react": "^18.2.0",
     "react-reconciler": "^0.28.0",
     "rpc-websockets": "^7.5.0",
-    "websocket": "^1.0.34"
+    "websocket": "^1.0.34",
+    "blast-utils": "workspace:*"
   },
   "devDependencies": {
-    "@types/debug": "^4.1.7",
-    "@types/fs-extra": "^9.0.13",
-    "@types/jest": "^29.2.4",
     "@types/lodash": "^4.14.191",
-    "@types/node": "^18.8.3",
     "@types/react-reconciler": "^0.28.0",
-    "@types/webpack": "^5.28.0",
     "@types/websocket": "^1.0.5",
-    "@typescript-eslint/eslint-plugin": "^5.46.1",
-    "@typescript-eslint/parser": "^5.47.0",
-    "cross-env": "^7.0.3",
-    "eslint": "^8.30.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-import-resolver-alias": "^1.1.2",
-    "eslint-import-resolver-typescript": "^3.5.2",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-react": "^7.31.11",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "jest": "^29.3.1",
-    "nanoid": "^4.0.0",
-    "nodemon": "^2.0.20",
-    "prettier": "2.8.1",
-    "react-devtools": "^4.27.1",
-    "ts-loader": "^9.4.2",
-    "ts-node": "^10.9.1",
-    "tslib": "^2.4.1",
-    "typescript": "^4.9.4",
-    "webpack": "5",
-    "webpack-cli": "^5.0.1"
+    "react-devtools": "^4.27.1"
   },
   "jest": {
     "testEnvironment": "node",
