+++ packages/blast-renderer/src/index.ts
@@ -1,2 +1,2 @@
-export * from "./renderer/render";
-export * from "./renderer/reconciler";
+export * from "./renderer";
+export { runServer } from "./server";
