--- packages/blast-renderer/src/raycast/types.ts
@@ -1,3 +1,0 @@
-import * as ComponentTypes from "../renderer/elements/types";
-
-export { ComponentTypes };
