+++ packages/blast-renderer/src/renderer/elements/index.ts
@@ -1,7 +1,10 @@
-import { createDebug } from "../../utils/debug";
+import { createDebug } from "blast-utils";
 
 export const debug = createDebug("blast:elements:index");
 
 import createElement from "./createElement";
 
 export default createElement;
+
+export * from "./elements";
+export * as ElementTypes from "./types";
