+++ packages/blast-renderer/src/renderer/run.ts
@@ -1,10 +1,10 @@
 import React from "react";
 import type { Server } from "rpc-websockets";
 
-import { App } from "../examples/App";
+// import { App } from "../../../blast-runtime/App";
 
 import { render } from "./render";
 
-export function run(Command: React.FC, { server }: { server: Server }) {
-  render(React.createElement(App, { server, Command }), server);
-}
+// export function run(Command: React.FC, { server }: { server: Server }) {
+//   render(React.createElement(App, { server, Command }), server);
+// }
