+++ packages/blast-runtime/index.ts
@@ -1,7 +1,7 @@
 import Command from "../../../../examples/todo-list/src/index";
 import { run } from "../renderer/run";
 import { runServer } from "../server";
-import { connect } from "../utils/preloadDevtool";
+import { connect } from "../utils/connectDevtools";
 
 connect();
